src/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["cardene"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
django-cloudinary-storage = "^0.3.0"
pillow = "^9.4.0"
django-filter = "^22.1"
psycopg2 = "^2.9.5"
dj-database-url = "^1.2.0"
mysqlclient = "^2.1.1"
pythonanywhere = "^0.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
